prettier + solhint

+++ contracts/interfaces/IUniswapV3Callee.sol
@@ -2,5 +2,10 @@
 pragma solidity >=0.5.0;
 
 interface IUniswapV3Callee {
-    function uniswapV3Call(address sender, uint amount0, uint amount1, bytes calldata data) external;
+    function uniswapV3Call(
+        address sender,
+        uint256 amount0,
+        uint256 amount1,
+        bytes calldata data
+    ) external;
 }
