+++ contracts/test/UniswapV3PairTest.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 pragma solidity =0.6.11;
 
-import "../interfaces/IUniswapV3Pair.sol";
+import '../interfaces/IUniswapV3Pair.sol';
 
 contract UniswapV3PairTest {
     IUniswapV3Pair pair;
@@ -10,14 +10,14 @@ contract UniswapV3PairTest {
         pair = IUniswapV3Pair(pair_);
     }
 
-    function getGasCostOfGetFee() public view returns (uint) {
-        uint gasBefore = gasleft();
+    function getGasCostOfGetFee() public view returns (uint256) {
+        uint256 gasBefore = gasleft();
         pair.getFee();
         return gasBefore - gasleft();
     }
 
-    function getGasCostOfGetVirtualSupply() public view returns (uint) {
-        uint gasBefore = gasleft();
+    function getGasCostOfGetVirtualSupply() public view returns (uint256) {
+        uint256 gasBefore = gasleft();
         pair.getVirtualSupply();
         return gasBefore - gasleft();
     }
